Change internal event naming scheme
