+++ modality-probe-cli/tests/fixtures/test-log.jsonl
