+++ modality-probe-cli/tests/stable_uuid.rs
@@ -1,3 +1,5 @@
+#![deny(warnings)]
+
 use std::fs::{self, File};
 use std::io::Write;
