--- tests/fixtures/test-component/events.csv
@@ -1,10 +1,0 @@
-component_id,id,name,description,tags,type_hint,file,line
-19894182-6b45-4e2e-8bc7-415f01096287,1073741822,INTERNAL_EVENT_PRODUCED_EXTERNAL_REPORT,The probe produced a log report for transmission to the backend for external analysis,internal,,,
-19894182-6b45-4e2e-8bc7-415f01096287,1073741821,INTERNAL_EVENT_LOG_ITEMS_MISSED,"Some log items were overwritten without successfully getting reported to the collector, the number of missed entries is storedin the payload",internal,u32,,
-19894182-6b45-4e2e-8bc7-415f01096287,1073741820,INTERNAL_EVENT_LOGICAL_CLOCK_OVERFLOWED,A logical clock's count reached the maximum trackable value,internal,,,
-19894182-6b45-4e2e-8bc7-415f01096287,1073741819,INTERNAL_EVENT_NUM_CLOCKS_OVERFLOWED,The probe did not have enough memory reserved to store enough logical clocks to track all of the unique neighbors that attempt to communicate with it,internal,,,
-19894182-6b45-4e2e-8bc7-415f01096287,1073741818,EVENT_INSUFFICIENT_REPORT_BUFFER_SIZE,The report destination buffer is too small to fit a header and/or the frontier clocks,internal,,,
-19894182-6b45-4e2e-8bc7-415f01096287,1,TOP_OF_THE_LOOP,At the top of the loop,example;my-tag,,examples/event-recording.rs,36
-19894182-6b45-4e2e-8bc7-415f01096287,2,MOD10_CONDITION_EVENT,Loop counter % 10 event,EXPECTATION;example,u32,examples/event-recording.rs,44
-19894182-6b45-4e2e-8bc7-415f01096287,3,LOOP_COUNTER_EVENT,Loop counter event happened,,,examples/event-recording.rs,54
-19894182-6b45-4e2e-8bc7-415f01096287,4,REPORT_CREATED,Report created,another tag,u32,examples/event-recording.rs,68
